osa 4 tehty

+++ part_4/controllers/blogs.js
@@ -1,28 +1,51 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
+  const token = request.token
+  const body = request.body
+
   try {
-    const user = await User.findOne({})
-    const body = request.body
-    body.user = user.id
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog(body)
+
+    blog.user = user.id
     const result = await blog.save()
-    user.blogs = user.blogs.concat(result._id)
-    user.blogs = response.status(201).json(result)
+    user.blogs = user.blogs.concat(result.id)
+    await user.save()
+
+    response.json(result.toJSON())
   } catch (exception) {
     next(exception)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = request.token
+
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    console.log(blog)
+    const userId = blog.user.toString()
+    if (decodedToken.id !== userId) {
+      return response.status(401).json({ error: 'not authenticated' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
